+++ test/form/samples/undefined-var/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'treeshakes undefined variable literals'
+};