+++ test/form/samples/undefined-var/main.js
@@ -1,0 +1,16 @@
+var x;
+var y = undefined;
+var z;
+if (x)
+	console.log('yes');
+if (!x)
+	console.log('no');
+if (y)
+	console.log('yes');
+if (!y)
+	console.log('no');
+if (z)
+	console.log('yes');
+if (!z)
+	console.log('no');
+z = 1;