+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # rollup changelog
 
-## 1.0.0
+## 0.61.0
 *unreleased*
+* Handle undefined values when evaluating conditionals ([#2264](https://github.com/rollup/rollup/pull/2264))
 * Handle known undefined properties when evaluating conditionals ([#2265](https://github.com/rollup/rollup/pull/2265))
 * Access watch events via the plugin context ([#2261](https://github.com/rollup/rollup/pull/2261))